# Desktop: Improve Webpack resolve configuration