@dmsnell 2017-03-31
what a great find @spen!

do you think we could build with and without this PR applied and compare the output JavaScript? I wonder if they are character-for-character identical or not. are there any size differences?