@spen 2017-03-31
I remember the Webpack reports reading the same, but they do do some rounding - if I remember right the bundles were something like 5.72mb before and after, comparing with more fidelity would certainly be worthwhile.
I'll take a closer over the weekend :)