@spen 2017-04-03
Confirmed! Before === after :)

<img width="186" alt="screen shot 2017-04-03 at 22 50 53" src="https://cloud.githubusercontent.com/assets/4335450/24633369/153db496-18c0-11e7-9eb7-d2863db2a0e9.png">

Diffing in FileMerge also show no differences 👍