@dmsnell 2017-04-03
This is great @spen - a zero-diff means we should have high confidence here. Thanks for performing that check. I'm comfortable merging this but I would strongly recommend running it by @gwwar and @ockham and anyone else they recommend so that they can chime in and be aware of the change.

Thanks again!