@gwwar 2017-04-03
I love faster builds. Changes here seem fine to merge, especially since the target branch is another feature branch. Local dev and desktop seem to spin up great. @spen was there more info on setting the node path? `NODE_PATH=../server:../client webpack --config webpack.config.js`

Also summoning @blowery and @aduth if they had more insights.