@spen 2017-04-03
> was there more info on setting the node path? `NODE_PATH=../server:../client webpack --config webpack.config.js`

Sure thing! :)

In `wp-desktop`this short conversation is around setting up a similar config to `NODE_PATH`: https://github.com/Automattic/wp-desktop/pull/275/commits/edf788111b0a3b64220ddbc8f4b1871f778ae3c1#r105904602

In the target branch I've edited this to look up one level now that the relationship between Calypso and desktop have flipped around: https://github.com/Automattic/wp-calypso/pull/12430/files#diff-b1cd36ac3f156ef4dc61bff0f03cf429R98