@dmsnell 2017-04-07
@spen if the `NODE_PATH` isn't essential to Calypso or this PR and isn't related to the changes, let's :shipit: - I think this is a simple and straightforward win here