@spen 2017-04-07
I'd say it's unrelated (at least directly).
I'm having some issues with either the `NODE_PATH` or the webpack config actually - but I'll make sure that's mentioned on the feature branch as I'm 99% sure that this change-set isn't effected by it :)