@dmsnell 2017-04-08
> I'm having some issues with either the NODE_PATH or the webpack config actually - but I'll make sure that's mentioned on the feature branch

can you help explain that? basically I want to know if there's any reason we shouldn't go ahead and merge this in Calypso