@spen 2017-04-08
Ahh, this PR is based off the desktop merge feature branch and only effects the `desktop/webpack.config.js` - the [root webpack resolver config](https://github.com/Automattic/wp-calypso/blob/master/webpack.config.js#L72-L80) for Calypso is already a-ok :)